+++ src/components/posts/article/fixed-scroll-to-top-button.tsx
@@ -34,17 +34,15 @@ export function FixedScrollToTopButton() {
   }, []);
 
   return (
-    <div className="lg:hidden">
+    <div>
       <div ref={targetRef} />
-      <div
+      <ScrollToTopButton
         className={cn(
           // 글 본문 코드블록에 가려지는 문제 해결을 위해 z-index 5로 설정. sticky header에는 가려지도록 헤더의 z-index 10보다는 작게 설정
-          "fixed bottom-4 right-4 bg-background rounded-md z-[5]",
-          showButton ? "block" : "hidden",
+          "fixed bottom-4 right-4 lg:bottom-6 lg:right-6 z-[5]",
+          showButton ? "inline-flex" : "hidden",
         )}
-      >
-        <ScrollToTopButton />
-      </div>
+      />
     </div>
   );
 }
