+++ src/components/posts/article/scroll-to-top-button.tsx
@@ -2,8 +2,13 @@
 
 import { ArrowUpToLine } from "lucide-react";
 import { Button } from "@/components/ui/button";
+import { cn } from "@/utils/cn";
 
-export function ScrollToTopButton() {
+type Props = {
+  className?: string;
+};
+
+export function ScrollToTopButton({ className }: Props) {
   const handleClickScrollToTopButton = () => {
     window.scrollTo({ top: 0 });
   };
@@ -14,6 +19,10 @@ export function ScrollToTopButton() {
       onClick={handleClickScrollToTopButton}
       variant="outline"
       size="icon"
+      className={cn(
+        "lg:size-10 rounded-full bg-brand dark:bg-brand text-brand-foreground dark:text-brand-foreground border-0 hover:bg-brand hover:text-brand-foreground hover:dark:bg-brand hover:dark:text-brand-foreground",
+        className,
+      )}
     >
       <ArrowUpToLine className="size-5" />
     </Button>
