+++ src/schema/posts.ts
@@ -21,7 +21,6 @@ export type PostContentHeading = {
 
 export type Post = PostMatter & {
   id: string;
-  rawContent: string; // 읽어들인 .md, .mdx 파일 내용
-  bundledContent: string; // mdx-bundler에 의해 처리된 파일 내용
+  mdxContent: string;
   headings: PostContentHeading[];
 };
